+++ assets/js/editor/Editor.tsx
@@ -168,6 +168,18 @@ export default function Editor({
     (editor: any, monaco: typeof Monaco) => {
       setMonaco(monaco);
 
+      editor.addCommand(
+        // https://microsoft.github.io/monaco-editor/typedoc/classes/KeyMod.html
+        // https://microsoft.github.io/monaco-editor/typedoc/enums/KeyCode.html
+        monaco.KeyMod.CtrlCmd | monaco.KeyCode.Enter,
+        function () {
+          const manual_run_form = document.getElementById('manual_run_form')!;
+          manual_run_form.dispatchEvent(
+            new Event('submit', { bubbles: true, cancelable: true })
+          );
+        }
+      );
+
       monaco.languages.typescript.javascriptDefaults.setCompilerOptions({
         // This seems to be needed to track the modules in d.ts files
         allowNonTsExtensions: true,